e2e: refactor the provisioner code Fixes: #490
