+++ test/provisioner/kustomize.go
@@ -1,4 +1,4 @@
-package e2e
+package provisioner
 
 import (
 	"bytes"
